requirements.txt
Django>=6.0.1
djangorestframework>=3.16.1
djangorestframework-simplejwt>=5.5.1
psycopg2-binary>=2.9.11
celery>=5.6.2
redis>=7.1.0
gunicorn>=24.1.1
whitenoise>=6.6.0
pytest>=9.0.2
pytest-django>=4.11.1
django-cors-headers>=4.9.0
coreapi>=2.3.3
openai>=2.15.0